backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
mysqlclient==2.2.0
django-cors-headers==4.3.1
django-filter==23.3
python-dotenv==1.0.0
pytest==7.4.2
pytest-django==4.7.0
